requirements.txt
greenlet==3.0.1
playwright==1.40.0
pyee==11.0.1
pyscopg2==66.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
